Peripherals: Fix compiler warnings on Windows

+++ xbmc/peripherals/addons/PeripheralAddon.h
@@ -62,9 +62,9 @@ namespace PERIPHERALS
     PeripheralPtr GetPeripheral(unsigned int index) const;
     PeripheralPtr GetByPath(const std::string &strPath) const;
     bool         SupportsFeature(PeripheralFeature feature) const;
-    int          GetPeripheralsWithFeature(PeripheralVector &results, const PeripheralFeature feature) const;
-    size_t       GetNumberOfPeripherals(void) const;
-    size_t       GetNumberOfPeripheralsWithId(const int iVendorId, const int iProductId) const;
+    unsigned int GetPeripheralsWithFeature(PeripheralVector &results, const PeripheralFeature feature) const;
+    unsigned int GetNumberOfPeripherals(void) const;
+    unsigned int GetNumberOfPeripheralsWithId(const int iVendorId, const int iProductId) const;
     void         GetDirectory(const std::string &strPath, CFileItemList &items) const;
 
     /** @name Peripheral add-on methods */
